renamed modifiers into find proxies which is more meaningful
